(fleet) rename services => packages
This PR renames services to packages as it better reflects what this package has become.
