+++ pkg/fleet/installer/installer.go
@@ -21,8 +21,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/paths"
 
 	fleetEnv "github.com/DataDog/datadog-agent/pkg/fleet/env"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/repository"
-	"github.com/DataDog/datadog-agent/pkg/fleet/installer/service"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/db"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/oci"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -476,7 +476,7 @@ func (i *installerImpl) InstrumentAPMInjector(ctx context.Context, method string
 		return fmt.Errorf("APM injector is not installed")
 	}
 
-	err = service.InstrumentAPMInjector(ctx, method)
+	err = packages.InstrumentAPMInjector(ctx, method)
 	if err != nil {
 		return fmt.Errorf("could not instrument APM: %w", err)
 	}
@@ -496,7 +496,7 @@ func (i *installerImpl) UninstrumentAPMInjector(ctx context.Context, method stri
 		return fmt.Errorf("APM injector is not installed")
 	}
 
-	err = service.UninstrumentAPMInjector(ctx, method)
+	err = packages.UninstrumentAPMInjector(ctx, method)
 	if err != nil {
 		return fmt.Errorf("could not instrument APM: %w", err)
 	}
@@ -539,9 +539,9 @@ func (i *installerImpl) Close() error {
 func (i *installerImpl) startExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return service.StartAgentExperiment(ctx)
+		return packages.StartAgentExperiment(ctx)
 	case packageDatadogInstaller:
-		return service.StartInstallerExperiment(ctx)
+		return packages.StartInstallerExperiment(ctx)
 	default:
 		return nil
 	}
@@ -550,9 +550,9 @@ func (i *installerImpl) startExperiment(ctx context.Context, pkg string) error {
 func (i *installerImpl) stopExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return service.StopAgentExperiment(ctx)
+		return packages.StopAgentExperiment(ctx)
 	case packageDatadogInstaller:
-		return service.StopInstallerExperiment(ctx)
+		return packages.StopInstallerExperiment(ctx)
 	default:
 		return nil
 	}
@@ -561,9 +561,9 @@ func (i *installerImpl) stopExperiment(ctx context.Context, pkg string) error {
 func (i *installerImpl) promoteExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return service.PromoteAgentExperiment(ctx)
+		return packages.PromoteAgentExperiment(ctx)
 	case packageDatadogInstaller:
-		return service.PromoteInstallerExperiment(ctx)
+		return packages.PromoteInstallerExperiment(ctx)
 	default:
 		return nil
 	}
@@ -572,11 +572,11 @@ func (i *installerImpl) promoteExperiment(ctx context.Context, pkg string) error
 func (i *installerImpl) setupPackage(ctx context.Context, pkg string, args []string) error {
 	switch pkg {
 	case packageDatadogInstaller:
-		return service.SetupInstaller(ctx)
+		return packages.SetupInstaller(ctx)
 	case packageDatadogAgent:
-		return service.SetupAgent(ctx, args)
+		return packages.SetupAgent(ctx, args)
 	case packageAPMInjector:
-		return service.SetupAPMInjector(ctx)
+		return packages.SetupAPMInjector(ctx)
 	default:
 		return nil
 	}
@@ -585,11 +585,11 @@ func (i *installerImpl) setupPackage(ctx context.Context, pkg string, args []str
 func (i *installerImpl) removePackage(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return service.RemoveAgent(ctx)
+		return packages.RemoveAgent(ctx)
 	case packageAPMInjector:
-		return service.RemoveAPMInjector(ctx)
+		return packages.RemoveAPMInjector(ctx)
 	case packageDatadogInstaller:
-		return service.RemoveInstaller(ctx)
+		return packages.RemoveInstaller(ctx)
 	default:
 		return nil
 	}
