+++ pkg/fleet/installer/packages/apm_inject_windows.go
@@ -5,8 +5,8 @@
 
 //go:build windows
 
-// Package service provides a way to interact with os services
-package service
+// package packages provides a way to interact with os services
+package packages
 
 import "context"
