+++ pkg/fleet/installer/packages/datadog_installer.go
@@ -5,7 +5,7 @@
 
 //go:build !windows
 
-package service
+package packages
 
 import (
 	"context"
