+++ pkg/fleet/installer/packages/embedded/datadog-installer.service
