+++ pkg/fleet/installer/packages/embedded/dd-cleanup
