+++ pkg/fleet/installer/packages/embedded/embed.go
