+++ pkg/fleet/installer/packages/systemd_windows.go
@@ -5,8 +5,8 @@
 
 //go:build windows
 
-// Package service provides a way to interact with os services
-package service
+// package packages provides a way to interact with os services
+package packages
 
 // IsSystemdRunning noop
 func IsSystemdRunning() (running bool, err error) {
